[comment crud]

+++ project02/src/main/resources/static/js/detail_main.js
@@ -50,6 +50,12 @@ function addDetail(id, username, title, contents, modifiedAt) {
 
 function showHide() {
     $('#edit').on('click', function () {
+        let username = $('.username').text()
+        let cur_username = $('.username01').text()
+        if (cur_username != username){
+            alert("자신이 작성한 글만 수정이 가능합니다!")
+            return;
+        }
         $('.detail').hide()
         $('.detail__edit').show()
         let title = $('.title').text()
@@ -72,6 +78,7 @@ function editArticle() {
     let title = $('.detail-input').val();
     let contents = $('.detail-textarea').val();
     let username = $('.username').text()
+
     if (title == '') {
         alert("수정하실 제목을 적어 주세요!")
         return
@@ -94,7 +101,15 @@ function editArticle() {
 }
 
 function deleteArticle() {
+
     $('#delete').on('click', function () {
+        let username = $('.username').text()
+        let cur_username = $('.username01').text()
+
+        if (cur_username != username){
+            alert("자신이 작성한 글만 삭제가 가능합니다!")
+            return;
+        }
         let id = location.search.split('=')[1]
         $.ajax({
             type: "DELETE",
@@ -109,9 +124,9 @@ function deleteArticle() {
 
 function create_comment() {
     let article_id = location.search.split('=')[1]
-    let username = $('.username').text();
+    let name = $('.username01').text();
     let contents = $('.post__comment-textarea').val();
-    if (!username || $('.link-signup').text() == '로그인 하러 가기') {
+    if (!name || $('.link-signup').text() == '로그인 하러 가기') {
         alert("로그인을 하셔야 댓글을 달수 있습니다!")
         return;
     }
@@ -119,7 +134,7 @@ function create_comment() {
         alert("댓글을 적어주세요!!")
         return
     }
-    let data = {'article_id':article_id, 'username':username, 'contents':contents};
+    let data = {'article_id':article_id, 'username':name, 'contents':contents};
 
     $.ajax({
         type:'POST',
@@ -127,6 +142,7 @@ function create_comment() {
         contentType: 'application/json',
         data: JSON.stringify(data),
         success: function (response){
+            console.log(data)
             alert('댓글이 성공적으로 작성되었습니다!')
             window.location.reload();
         }
@@ -142,10 +158,10 @@ function getComment(id) {
         url: `/api/comment/${index}`,
         success: function (response) {
             for(let i=0; i< response.length; i++){
-                console.log(response['comments'][i])
-                // let comment = response[i];
-                // let tempHtml = addComments(comment)
-                // $('.comment__card-box').append(tempHtml);
+                console.log(response[i])
+                let comment = response[i];
+                let tempHtml = addComments(comment)
+                $('.comment__card-box').append(tempHtml);
             }
         }
     });
@@ -153,16 +169,98 @@ function getComment(id) {
 
 function addComments(comment){
     return `<div class="comment__card">
-                <div class="comment__card-header">
-                  ${comment.username}<span class="comment-date">${comment.modifiedAt}</span>
+            <div class="comment__card-header">
+              <span id="${comment.id}-comment_user">${comment.username}</span><span class="comment-date">${comment.modifiedAt}</span>
+            </div>
+            <div class="comment__card-body">
+              <div class="edit-comment-wrap">
+                <div id="${comment.id}-comment" class="comment">${comment.contents}</div>
+                <textarea
+                  class="edit-comment-textarea"
+                  id="${comment.id}-edit-comment-textarea"
+                  name=""
+                  id=""
+                  cols="30"
+                  rows="10"
+                ></textarea>
+              </div>
+              <div class="comment-btn">
+                <div class="befor" id="${comment.id}-befor">
+                  <i class="far fa-edit" id="${comment.id}-edit-btn" onclick="edit_start_comment(${comment.id})" ></i>
+                  <i class="fas fa-trash-alt" id="${comment.id}-delete-btn" onclick="delete_comment(${comment.id})"></i>
                 </div>
-                <div class="comment__card-body">
-                  <div class="comment">${comment.contents}</div>
-                  <div class="comment-btn">
-                    <i class="far fa-edit"></i>
-                    <i class="fas fa-trash-alt"></i>
-                  </div>
+                <div class="after" id="${comment.id}-after">
+                  <i class="fas fa-check"  id="${comment.id}-check-btn" onclick="edit_comment(${comment.id})" ></i>
+                  <i class="fas fa-times" id="${comment.id}-cancel-btn" onclick="edit_end_comment(${comment.id})" ></i>
                 </div>
-              </div>`;
+              </div>
+            </div>
+          </div>`;
 
+}
+
+function edit_start_comment(id){
+    let comment_user = $(`#${id}-comment_user`).text()
+    let cur_username = $('.username01').text()
+
+    if (comment_user != cur_username){
+        alert("본인의 댓글만 수정 하실수 있습니다.")
+        return;
+    }
+    $(`#${id}-befor`).hide()
+    $(`#${id}-after`).show()
+    $(`#${id}-edit-comment-textarea`).show()
+    $(`#${id}-comment`).hide()
+    $(`#${id}-edit-comment-textarea`).val($(`#${id}-comment`).text())
+
+}
+
+function edit_comment(id){
+    let article_id = location.search.split('=')[1]
+    let username = $('.username01').text()
+    let contents = $(`#${id}-edit-comment-textarea`).val()
+    if(contents==''){
+        alert('수정하실 내용을 작성 해주세요!')
+        return;
+    }
+    let data = {'article_id':article_id,'username':username,'contents':contents}
+
+    $.ajax({
+        type:'PUT',
+        url:`/api/comments/${id}`,
+        contentType:'application/json',
+        data:JSON.stringify(data),
+        success:function (response){
+            alert('댓글이 수정 되었습니다.')
+            window.location.reload();
+        }
+    })
+}
+
+
+function edit_end_comment(id){
+    $(`#${id}-befor`).show()
+    $(`#${id}-after`).hide()
+
+    $(`#${id}-edit-comment-textarea`).hide()
+    $(`#${id}-comment`).show()
+    $(`#${id}-edit-comment-textarea`).val('')
+}
+
+function delete_comment(id){
+    let comment_user = $(`#${id}-comment_user`).text()
+    let cur_username = $('.username01').text()
+
+    if (comment_user != cur_username){
+        alert("본인의 댓글만 삭제하실수 있습니다.")
+        return;
+    }
+    $.ajax({
+        type:'DELETE',
+        url:`api/comments/${id}`,
+        success: function (response){
+            alert('댓글이 삭제 되었습니다.')
+            window.location.reload();
+        }
+    })
 }
